VxAdmin: Move logs to Settings screen
Seemed weird to have a whole screen for logs with just one button on it.
Since other apps have log export on the settings screen, move it there
for consistency.
