+++ apps/admin/frontend/src/components/navigation_screen.tsx
@@ -37,13 +37,11 @@ interface Props {
 const SYSTEM_ADMIN_NAV_ITEMS: readonly NavItem[] = [
   { label: 'Election', routerPath: routerPaths.election },
   { label: 'Smartcards', routerPath: routerPaths.smartcards },
-  { label: 'Logs', routerPath: routerPaths.logs },
   { label: 'Settings', routerPath: routerPaths.settings },
 ];
 
 const SYSTEM_ADMIN_NAV_ITEMS_NO_ELECTION: readonly NavItem[] = [
   { label: 'Election', routerPath: routerPaths.election },
-  { label: 'Logs', routerPath: routerPaths.logs },
   { label: 'Settings', routerPath: routerPaths.settings },
 ];
