+++ apps/admin/frontend/src/screens/settings_screen.tsx
@@ -1,6 +1,7 @@
 import React, { useContext } from 'react';
 import {
   CurrentDateAndTime,
+  ExportLogsButton,
   H2,
   P,
   RebootToBiosButton,
@@ -12,15 +13,26 @@ import {
   isSystemAdministratorAuth,
 } from '@votingworks/utils';
 
+import type { LogsResultType } from '@votingworks/backend';
+import { err } from '@votingworks/basics';
 import { AppContext } from '../contexts/app_context';
 import { NavigationScreen } from '../components/navigation_screen';
 import { FormatUsbButton } from '../components/format_usb_modal';
-import { logOut } from '../api';
+import { exportLogsToUsb, logOut } from '../api';
 import { LiveCheckButton } from '../components/live_check_button';
 
 export function SettingsScreen(): JSX.Element {
-  const { auth, logger } = useContext(AppContext);
+  const { auth, logger, usbDriveStatus } = useContext(AppContext);
   const logOutMutation = logOut.useMutation();
+  const exportLogsToUsbMutation = exportLogsToUsb.useMutation();
+
+  async function doExportLogs(): Promise<LogsResultType> {
+    try {
+      return await exportLogsToUsbMutation.mutateAsync();
+    } catch (e) {
+      return err('copy-failed');
+    }
+  }
 
   return (
     <NavigationScreen title="Settings">
@@ -35,6 +47,13 @@ export function SettingsScreen(): JSX.Element {
       </P>
       {isSystemAdministratorAuth(auth) && (
         <React.Fragment>
+          <H2>Logs</H2>
+          <ExportLogsButton
+            usbDriveStatus={usbDriveStatus}
+            auth={auth}
+            logger={logger}
+            onExportLogs={doExportLogs}
+          />
           <H2>USB Formatting</H2>
           <FormatUsbButton />
           <H2>Software Update</H2>
